+++ src/actions/internal/setup-helpers/index.js
@@ -1,2 +1,3 @@
 export * from './execute-external-purge.js';
+export * from './execute-external-requirements-check.js';
 export * from './execute-external-setup.js';